services/scheduler/requirements.txt
flask
etcd3
protobuf == 3.19.0
chardet
pyyaml
kubernetes